fix(ci): GRA-386 (v3) clippy rust 1.97 lints — len_zero, clone_on_copy, unnecessary_cast, redundant_locals
6 lint fixes surfaced by operator-merged commits on main:
- porkchop.rs: grid.cells.len()>0 → !grid.cells.is_empty() (×2)
- porkchop.rs: planet_orbit.clone() → planet_orbit (×2, KeplerOrbit is Copy)
- visuals.rs: SCALING_FACTOR as f64 → SCALING_FACTOR (f64→f64 cast)
- transfer_planner.rs: redundant let target_entity = target_entity binding

+++ src/fleets/visuals.rs
@@ -1463,7 +1463,7 @@ mod tests {
         );
 
         // The polyline operates in Bevy render units (AU × SCALING_FACTOR).
-        let scale = SCALING_FACTOR as f64;
+        let scale = SCALING_FACTOR;
 
         // (1) Endpoint must be Jupiter's orbit (apoapsis = r_jupiter_au),
         //     not rotated past it by `mean_motion * elapsed` as the bug did.
